delete registro success

+++ routes/web.php
@@ -53,6 +53,6 @@ Auth::routes();
         Route::get('/monitoreo/modificar/{semana}','monitoreo@GetSemana');
         Route::post('/monitoreo/modificar/admin/habilitarreg','monitoreo@habilitarreg');
         Route::get('/monitoreo/eliminar/{semana}','monitoreo@Eliminar');
-        Route::delete('');
+        Route::delete('/admin/monitoreo/eliminar/{idregistro}','monitoreo@EliminarRegistro');
         Route::get('monitoreo/resultadosbacteriologicos','monitoreo@ResultadosBacteriologicos');
     });
